ai-model/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
scikit-learn==1.6.1
pandas==2.2.3
joblib==1.4.2
python-dotenv==1.0.1
Pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.20
